+++ tests/forge/expected/codec_zenoh_locator.rs
@@ -84,7 +84,7 @@ impl<'a> CodecZenohLocator<'a> {
         // `vle_encode_block` with the language-appropriate self/
         // struct prefix.
         {
-            let mut _vle = self.locator_len as u64;
+            let mut _vle = self.locator_len;
             while _vle >= 0x80 {
                 w.write_u8((_vle as u8 & 0x7F) | 0x80)?;
                 _vle >>= 7;
